requirements.txt
web3
flask
ccxt
python-dotenv
